Refactored date functions.
